[flink] minor changes
