+++ samoa-flink/src/main/resources/log4j.properties
@@ -1,25 +1,27 @@
 #/*
 # * #%L
 # * SAMOA
-# * %%
-# * Copyright (C) 2013 - 2015 Yahoo! Inc.
-# * %%
-# * Licensed under the Apache License, Version 2.0 (the "License");
-# * you may not use this file except in compliance with the License.
-# * You may obtain a copy of the License at
 # *
-# *      http://www.apache.org/licenses/LICENSE-2.0
+# %%
+# Copyright (C) 2013 - 2015 Yahoo! Inc.
 # *
-# * Unless required by applicable law or agreed to in writing, software
-# * distributed under the License is distributed on an "AS IS" BASIS,
-# * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# * See the License for the specific language governing permissions and
-# * limitations under the License.
-# * #L%
+# %%
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+#      http://www.apache.org/licenses/LICENSE-2.0
+# 
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# #L%
 # */
 
 
-log4j.rootLogger=ERROR,stdout
+log4j.rootLogger=DEBUG,stdout
 log4j.logger.com.endeca=INFO
 # Logger for crawl metrics
 log4j.logger.com.endeca.itl.web.metrics=INFO
